refactor : `PersistenceEnvironment` 리팩토링
- 패키지 core 로 이동
- `DmlGenerator` 를 직접 가지고 있게 변경
