+++ src/main/java/persistence/core/PersistenceEnvironment.java
@@ -1,14 +1,17 @@
-package persistence.dialect;
+package persistence.core;
 
+import persistence.dialect.Dialect;
 import persistence.sql.dml.DmlGenerator;
 
 import java.sql.Connection;
 
 public class PersistenceEnvironment {
     private final PersistenceEnvironmentStrategy strategy;
+    private final DmlGenerator dmlGenerator;
 
     public PersistenceEnvironment(final PersistenceEnvironmentStrategy strategy) {
         this.strategy = strategy;
+        this.dmlGenerator = new DmlGenerator(strategy.getDialect());
     }
 
 
@@ -21,6 +24,6 @@ public class PersistenceEnvironment {
     }
 
     public DmlGenerator getDmlGenerator() {
-        return this.strategy.getDmlGenerator();
+        return this.dmlGenerator;
     }
 }
