+++ src/test/java/persistence/ApplicationTest.java
@@ -8,9 +8,9 @@ import jdbc.RowMapper;
 import org.junit.jupiter.api.*;
 import persistence.core.EntityMetadata;
 import persistence.core.EntityMetadataProvider;
-import persistence.dialect.DefaultPersistenceEnvironmentStrategy;
+import persistence.core.DefaultPersistenceEnvironmentStrategy;
 import persistence.dialect.H2Dialect;
-import persistence.dialect.PersistenceEnvironment;
+import persistence.core.PersistenceEnvironment;
 import persistence.entity.EntityManager;
 import persistence.entity.SimpleEntityManager;
 import persistence.sql.ddl.DdlGenerator;
